size

+++ 0x00-hello_world/6-size.c
@@ -13,7 +13,7 @@ int main(void)
 	printf("sizeof a char: %lu byte(s)\n", sizeof(char));
 	printf("sizeof int: %lu byte(s)\n", sizeof(int));
 	printf("sizeof long int: %lu byte(s)\n", sizeof(long int));
-	printf("sizeof long long int: %lu byte(s)\n", sizeof(long int));
+	printf("sizeof long long int: %lu byte(s)\n", sizeof(long long int));
 	printf("sizeof float: %lu byte(s)\n", sizeof(float));
 	return (0);
 }
